+++ assets/javascripts/kitten/components/layout/desk-layout/stories.js
@@ -1,16 +1,25 @@
 import React from 'react'
 import { DeskLayout } from './index'
 import { DocsPage } from 'storybook/docs-page'
+import { createGlobalStyle } from 'styled-components'
 import {
   DeskMenu,
   Text,
   HeaderNav,
   KissKissBankBankLogo,
   KissKissLoadingAnimation,
+  HEADER_HEIGHT,
+  pxToRem,
 } from 'kitten'
 import { WithContributions } from './stories/with-contributions'
 import { WithProjects } from './stories/with-projects'
 import { WithForm } from './stories/with-form'
+
+const GlobalStyle = createGlobalStyle`
+  .k-HeaderNav--isStuck + * .k-DeskLayout__menu__inside {
+    top: ${pxToRem(HEADER_HEIGHT + 20)} !important;
+  }
+`
 
 export default {
   title: 'Layout/DeskLayout',
@@ -90,9 +99,9 @@ const VisitorMenu = props => {
 
 export const Default = args => (
   <div>
+    <GlobalStyle />
     <HeaderNav
       id="header_id"
-      isFixed={true}
       quickAccessProps={{
         href: '#mainContent',
         text: 'Aller au contenu principal',
@@ -200,9 +209,9 @@ export const Default = args => (
 
 export const WithContributionsContent = args => (
   <div>
+    <GlobalStyle />
     <HeaderNav
       id="header_id"
-      isFixed={true}
       quickAccessProps={{
         href: '#mainContent',
         text: 'Aller au contenu principal',
@@ -232,9 +241,9 @@ export const WithContributionsContent = args => (
 
 export const WithProjectsContent = args => (
   <div>
+    <GlobalStyle />
     <HeaderNav
       id="header_id"
-      isFixed={true}
       quickAccessProps={{
         href: '#mainContent',
         text: 'Aller au contenu principal',
@@ -264,9 +273,9 @@ export const WithProjectsContent = args => (
 
 export const WithFormContent = args => (
   <div>
+    <GlobalStyle />
     <HeaderNav
       id="header_id"
-      isFixed={true}
       quickAccessProps={{
         href: '#mainContent',
         text: 'Aller au contenu principal',
@@ -295,9 +304,9 @@ export const WithFormContent = args => (
 
 export const Loading = args => (
   <div>
+    <GlobalStyle />
     <HeaderNav
       id="header_id"
-      isFixed={true}
       quickAccessProps={{
         href: '#mainContent',
         text: 'Aller au contenu principal',